Move pytest imports to tests/ remove requirememnt
This will need updating in the feedstock.
