+++ fastparquet/benchmarks/columns.py
@@ -1,5 +1,4 @@
 from contextlib import contextmanager
-import datetime
 import os
 import numpy as np
 import pandas as pd
