+++ fastparquet/test/test_aroundtrips.py
@@ -13,8 +13,7 @@ import pytest
 import fastparquet
 from fastparquet import write
 from fastparquet.compression import compressions
-from fastparquet.test.util import sql, s3, TEST_DATA
-from fastparquet.util import tempdir, default_mkdirs
+from fastparquet.test.util import sql, s3, tempdir, TEST_DATA
 
 
 def test_map_array(sql):
