+++ requirements.txt
@@ -2,6 +2,5 @@ pandas
 numba>=0.28
 numpy>=1.11
 thrift>=0.10.0
-pytest
 six
 cython
